Coordinate color of file manager icon with color of folders
Makes use of the circular folder icons by replacing the file manager icon (previously always blue) with them. As such the ubuntu version now has an ubuntu-colored file manager icon, red version has a red icon, etc. Not blue.

Note: as of right now, the install script throws an error for the nord version, because the circle icons do not yet exist for this theme. It will leave the file manager icon as blue, and continue with the installation process as usual.